+++ lib/rpush/client/active_record/apns/feedback.rb
@@ -6,8 +6,10 @@ module Rpush
           self.table_name = 'rpush_feedback'
 
           if Rpush.attr_accessible_available?
-            attr_accessible :device_token, :failed_at, :app
+            attr_accessible :device_token, :failed_at, :app_id
           end
+
+          belongs_to :app, class_name: 'Rpush::Client::ActiveRecord::App'
 
           validates :device_token, presence: true
           validates :failed_at, presence: true