+++ lib/rpush/daemon/store/active_record.rb
@@ -119,7 +119,7 @@ module Rpush
         def create_apns_feedback(failed_at, device_token, app)
           with_database_reconnect_and_retry do
             Rpush::Client::ActiveRecord::Apns::Feedback.create!(failed_at: failed_at,
-                                                                device_token: device_token, app: app)
+                                                                device_token: device_token, app_id: app.id)
           end
         end